Remove prometheus by default
